pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.2",
]
